requirements.txt
pypdf 
pandas 
langchain 
openai 
streamlit 
python-dotenv
requests
pdf2image 
pytesseract 
pillow
PyYAML
#os
